+++ CHANGELOG.md
@@ -22,6 +22,7 @@ CHANGELOG
  * Add `MicroKernelTrait::getAllowedEnvs()` to list and enforce allowed values for `APP_ENV`
  * Change `caching.max_ttl` option to default to `86400` (24h) instead of `null`
  * Deprecate setting `caching.max_ttl` to `null`, use a positive integer instead
+ * Deprecate `senders` nesting level for messenger routing config; use string or a list of strings instead
 
 8.0
 ---